start_settings
создание папок и убрал лишние методы
