+++ iOS10_HW12_Kretov_Igor/Resources/Assets.xcassets/AccentColor.colorset/Contents.json
