+++ iOS10_HW12_Kretov_Igor/Resources/Assets.xcassets/AppIcon.appiconset/Contents.json
