+++ iOS10_HW12_Kretov_Igor/Resources/Assets.xcassets/Contents.json
