+++ iOS10_HW12_Kretov_Igor/Resources/Base.lproj/LaunchScreen.storyboard
