+++ iOS10_HW12_Kretov_Igor/Resources/Info.plist
