+++ iOS10_HW12_Kretov_Igor/Sources/App/AppDelegate.swift
@@ -1,0 +1,20 @@
+//
+//  AppDelegate.swift
+//  iOS10_HW12_Kretov_Igor
+//
+//  Created by Igor Kretov on 06.07.2023.
+//
+
+import UIKit
+
+@main
+class AppDelegate: UIResponder, UIApplicationDelegate {
+
+
+
+    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
+        return true
+    }
+
+
+
