+++ iOS10_HW12_Kretov_Igor/Sources/App/SceneDelegate.swift
@@ -1,0 +1,15 @@
+//
+//  SceneDelegate.swift
+//  iOS10_HW12_Kretov_Igor
+//
+//  Created by Igor Kretov on 06.07.2023.
+//
+
+import UIKit
+
+class SceneDelegate: UIResponder, UIWindowSceneDelegate {
+    var window: UIWindow?
+    func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
+        guard let _ = (scene as? UIWindowScene) else { return }
+    }
+}
