+++ iOS10_HW12_Kretov_Igor/Sources/ViewController.swift
@@ -11,7 +11,6 @@ class ViewController: UIViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
-        // Do any additional setup after loading the view.
     }
